get rid of bot.Channel proxy class
